requirements.txt
annotated-types==0.7.0
anthropic==0.43.0
anyio==4.8.0
boto3==1.36.1
botocore==1.36.1
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inspeqai==1.0.30
jiter==0.8.2
jmespath==1.0.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.1
